# app/swagger.yaml
swagger: "2.0"
info:
  title: Medical Records API
  description: API untuk mengelola data rekam medis pasien, pengguna, laporan, dan dokumen medis lainnya.
  version: 1.0.0
host: localhost:5000
basePath: /
schemes:
  - http
produces:
  - application/json
consumes:
  - application/json
securityDefinitions:
  bearerAuth:
    type: apiKey
    name: Authorization
    in: header
    description: "Masukkan token JWT dengan format: 'Bearer {token}'"
security:
  - bearerAuth: []
...
definitions:
...
  Laporan:
    type: object
    properties:
      id:
        type: integer
      patient_id:
        type: integer
      cppt_id:
        type: integer
      tanggal:
        type: string
        format: date-time
      user_id:
        type: integer
      subjective:
        type: string
      objective:
        type: string
      assessment:
        type: string
      plan:
        type: string
      keterangan:
        type: string
      dokter:
        type: string
      signature:
        type: string
      tindakan_lanjutan:
        type: string
      SDKI:
        type: string
      SLKI:
        type: string
      SIKI:
        type: string
...
  Query:
    type: object
    properties:
      query:
        type: string
        description: "Teks atau kueri untuk diproses oleh AI atau pencarian."
paths:
...
  /laporan/:
    get:
      tags:
        - Laporan
      summary: Mendapatkan semua laporan
      security: []
      responses:
        '200':
          description: Sukses
    post:
      tags:
        - Laporan
      summary: Membuat laporan baru dengan bantuan AI
      parameters:
        - in: body
          name: body
          required: true
          schema:
            type: object
            properties:
              query:
                type: string
              cppt_id:
                type: integer
              patient_id:
                type: integer
              perawat_id:
                type: integer
      responses:
        '201':
          description: Laporan berhasil dibuat
  /laporan/search:
    post:
      tags:
        - Laporan
      summary: Mencari dan menghasilkan data laporan potensial dengan AI
      security: []
      parameters:
        - in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/Query'
      responses:
        '200':
          description: Hasil pencarian
  /laporan/{id}:
    parameters:
      - name: id
        in: path
        required: true
        type: integer
    get:
      tags:
        - Laporan
      summary: Mendapatkan detail laporan
      security: []
      responses:
        '200':
          description: Sukses
    put:
      tags:
        - Laporan
      summary: Memperbarui laporan
      parameters:
        - in: body
          name: body
          schema:
            $ref: '#/definitions/Laporan'
      responses:
        '200':
          description: Laporan berhasil diperbarui
    delete:
      tags:
        - Laporan
      summary: Menghapus laporan
      responses:
        '200':
          description: Laporan berhasil dihapus
...